# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Delta-Neutral Vault API",
    "version": "0.1.0",
    "description": "Read API over the delta-neutral ERC-4626 vault running against LOCAL MOCK MARKETS (no real funds). Money is in human units (USD numbers, ETH numbers), ratios in basis points, rates as fractions. `blockTimestamp` / `ts` fields are chain time: the demo warps time to build history."
  },
...
  "paths": {
...
    "/vault/performance": {
      "get": {
        "summary": "Historical series + summary statistics",
        "tags": [
          "vault"
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "in": "query",
            "name": "days",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "properties": {
                    "points": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true,
                        "properties": {
                          "ts": {
                            "type": "integer"
                          },
                          "tvlUsd": {
                            "type": "number"
                          },
                          "sharePrice": {
                            "type": "number"
                          },
                          "ethPrice": {
                            "type": "number"
                          },
                          "netDeltaBps": {
                            "type": "number"
                          },
                          "leverage": {
                            "type": "number"
                          },
                          "drawdownBps": {
                            "type": "number"
                          },
                          "fundingRatePer8h": {
                            "type": "number"
                          },
                          "apy7d": {
                            "type": [
                              "null",
                              "number"
                            ]
                          },
                          "apy30d": {
                            "type": [
                              "null",
                              "number"
                            ]
                          }
                        }
                      }
                    },
                    "summary": {
                      "type": "object",
                      "additionalProperties": true,
                      "properties": {
                        "fromTs": {
                          "type": "integer"
                        },
                        "toTs": {
                          "type": "integer"
                        },
                        "periodReturn": {
                          "type": "number"
                        },
                        "annualizedReturn": {
                          "type": [
                            "null",
                            "number"
                          ]
                        },
                        "maxDrawdownBps": {
                          "type": "number"
                        },
                        "volatilityAnnual": {
                          "type": [
                            "null",
                            "number"
                          ]
                        },
                        "sharpe": {
                          "type": [
                            "null",
                            "number"
                          ]
                        },
                        "rebalances": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/rebalance/history": {
      "get": {
        "summary": "Executed rebalances (indexed events)",
        "tags": [
          "strategy"
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000
            },
            "in": "query",
            "name": "limit",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true,
                    "properties": {
                      "id": {
                        "type": "integer"
                      },
                      "ts": {
                        "type": "integer"
                      },
                      "blockNumber": {
                        "type": "integer"
                      },
                      "txHash": {
                        "type": "string"
                      },
                      "triggers": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "urgent": {
                        "type": "boolean"
                      },
                      "longUsdDelta": {
                        "type": "number"
                      },
                      "marginDelta": {
                        "type": "number"
                      },
                      "perpSizeChange": {
                        "type": "number"
                      },
                      "preDeltaBps": {
                        "type": "number"
                      },
                      "postDeltaBps": {
                        "type": "number"
                      },
                      "preLeverage": {
                        "type": "number"
                      },
                      "postLeverage": {
                        "type": "number"
                      },
                      "estimatedCostUsd": {
                        "type": "number"
                      },
                      "realizedCostUsd": {
                        "type": "number"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/alerts": {
      "get": {
        "summary": "Alerts (active first)",
        "tags": [
          "risk"
        ],
        "parameters": [
          {
            "schema": {
              "type": "boolean"
            },
            "in": "query",
            "name": "active",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true,
                    "properties": {
                      "id": {
                        "type": "integer"
                      },
                      "key": {
                        "type": "string"
                      },
                      "severity": {
                        "type": "string",
                        "description": "INFO | WARNING | CRITICAL"
                      },
                      "title": {
                        "type": "string"
                      },
                      "message": {
                        "type": "string"
                      },
                      "value": {
                        "type": [
                          "null",
                          "number"
                        ],
                        "description": "Observed value, in `unit`"
                      },
                      "threshold": {
                        "type": [
                          "null",
                          "number"
                        ],
                        "description": "Threshold crossed, in `unit`"
                      },
                      "unit": {
                        "type": [
                          "null",
                          "string"
                        ],
                        "description": "bps | apr (annualised rate in bps) | usd | null when the alert carries no number"
                      },
                      "openedAt": {
                        "type": "integer",
                        "description": "Chain time"
                      },
                      "resolvedAt": {
                        "type": [
                          "null",
                          "integer"
                        ],
                        "description": "Chain time"
                      },
                      "active": {
                        "type": "boolean"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/simulation": {
      "post": {
        "summary": "Run a scenario, custom scenario, Monte Carlo or optimizer; result is persisted",
        "tags": [
          "simulation"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "description": "scenario | custom | montecarlo | optimizer",
                    "enum": [
                      "scenario",
                      "custom",
                      "montecarlo",
                      "optimizer"
                    ]
                  },
                  "scenarioId": {
                    "type": "string",
                    "description": "A..J (type=scenario)"
                  },
                  "priceMovePct": {
                    "type": "number",
                    "description": "type=custom (required): ETH move in percent, e.g. -30"
                  },
                  "moveDays": {
                    "type": "number",
                    "description": "type=custom (required): days over which the move happens"
                  },
                  "horizonDays": {
                    "type": "number",
                    "description": "type=custom (required): total simulated days"
                  },
                  "fundingRatePer8h": {
                    "type": "number"
                  },
                  "usdcUtilization": {
                    "type": "number"
                  },
                  "wethUtilization": {
                    "type": "number"
                  },
                  "oracleOutageDays": {
                    "type": "number"
                  },
                  "liquidityMultiplier": {
                    "type": "number"
                  },
                  "leverageBps": {
                    "type": "number"
                  },
                  "reserveBps": {
                    "type": "number"
                  },
                  "paths": {
                    "type": "integer",
                    "description": "type=montecarlo, max 500"
                  },
                  "days": {
                    "type": "integer",
                    "description": "type=montecarlo, max 730"
                  },
                  "seed": {
                    "type": "integer"
                  },
                  "tvl": {
                    "type": "number"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "type": {
                      "type": "string"
                    },
                    "result": {
                      "type": "object",
                      "additionalProperties": true,
                      "properties": {}
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/simulation/{id}": {
      "get": {
        "summary": "Fetch a stored simulation",
        "tags": [
          "simulation"
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "createdAt": {
                      "type": "string",
                      "description": "Wall-clock time the simulation ran (ISO 8601) - not chain time"
                    },
                    "kind": {
                      "type": "string"
                    },
                    "input": {
                      "type": "object",
                      "additionalProperties": true,
                      "properties": {}
                    },
                    "result": {
                      "type": "object",
                      "additionalProperties": true,
                      "properties": {}
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
...
}